src/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
py-bcrypt==0.4
requests==2.28.1
SQLAlchemy==1.4.42
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.15.0